Refactor: Routing, DB Config
