+++ src/routes/employeeRoutes.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import { getAllEmployees, getEmployeeById, createEmployee, updateEmployee, deleteEmployee } from '../controllers/employee/employeesControllers';
+import { getAllEmployees, getEmployeeById, createEmployee, updateEmployee, deleteEmployee, getEmployeePipelineById, getEmployeeBenchById, getEmployeeBillingById } from '../controllers/employeesController';
 
 const router = express.Router();
 
@@ -8,5 +8,9 @@ router.get('/employees/:id', getEmployeeById);
 router.post('/employees', createEmployee);
 router.patch('/employees/:id', updateEmployee);
 router.delete('/employees/:id', deleteEmployee);
+router.get('/employees/pipeline/:id', getEmployeePipelineById);
+router.get('/employees/bench/:id', getEmployeeBenchById);
+router.get('/employees/billing/:id', getEmployeeBillingById);
 
 export default router;
+
