+++ tsconfig.json
@@ -27,7 +27,7 @@
       /* Modules */
       "module": "commonjs",                                /* Specify what module code is generated. */
       // "rootDir": "./",                                  /* Specify the root folder within your source files. */
-      "moduleResolution": "node",                     /* Specify how TypeScript looks up a file from a given module specifier. */
+      "moduleResolution": "node",                         /* Specify how TypeScript looks up a file from a given module specifier. */
       // "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
       // "paths": {},                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
       // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
